お気に入り動画関連のテストをフォルダ分け
