+++ backend/tests/Feature/FavoriteVideo/ChildFolder/FetchChildFoldersTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\FavoriteVideo\ChildFolder;
 
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
-use App\Models\User;
 
 class FetchChildFoldersTest extends TestCase
 {
