+++ backend/tests/Feature/FavoriteVideo/ChildFolder/StoreChildFolderTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\FavoriteVideo\ChildFolder;
 
 use App\Models\ChildFolder;
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
-use App\Models\User;
 
 class StoreChildFolderTest extends TestCase
 {
