+++ backend/tests/Feature/FavoriteVideo/ChildFolder/UpdateChildFolderTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\FavoriteVideo\ChildFolder;
 
 use App\Models\ChildFolder;
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Testing\TestResponse;
 use Tests\TestCase;
-use App\Models\User;
 
 class UpdateChildFolderTest extends TestCase
 {
