+++ backend/tests/Feature/FavoriteVideo/DeleteFavoriteVideoTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\FavoriteVideo;
 
 use App\Models\FavoriteVideo;
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
-use App\Models\User;
 
 class DeleteFavoriteVideoTest extends TestCase
 {
