+++ backend/tests/Feature/FavoriteVideo/FetchFavoriteVideosTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\FavoriteVideo;
 
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
-use App\Models\User;
 
 class FetchFavoriteVideosTest extends TestCase
 {
