+++ backend/tests/Feature/FavoriteVideo/GrandchildFolder/DeleteGrandchildFolderTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\FavoriteVideo\GrandchildFolder;
 
 use App\Models\GrandchildFolder;
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
-use App\Models\User;
 
 class DeleteGrandchildFolderTest extends TestCase
 {
