+++ backend/tests/Feature/FavoriteVideo/GrandchildFolder/FetchGrandchildFoldersTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\FavoriteVideo\GrandchildFolder;
 
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
-use App\Models\User;
 
 class FetchGrandchildFoldersTest extends TestCase
 {
