+++ backend/tests/Feature/FavoriteVideo/GrandchildFolder/RegisterByGrandchildFolderTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\FavoriteVideo\GrandchildFolder;
 
 use App\Models\FavoriteVideo;
 use App\Models\GrandchildFolder;
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Facades\Auth;
 use Tests\TestCase;
-use App\Models\User;
 
 class RegisterByGrandchildFolderTest extends TestCase
 {
