+++ backend/tests/Feature/FavoriteVideo/GrandchildFolder/UpdateGrandchildFolderTest.php
@@ -1,13 +1,12 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\FavoriteVideo\GrandchildFolder;
 
-use App\Models\ChildFolder;
 use App\Models\GrandchildFolder;
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Testing\TestResponse;
 use Tests\TestCase;
-use App\Models\User;
 
 class UpdateGrandchildFolderTest extends TestCase
 {
