+++ backend/tests/Feature/FavoriteVideo/ParentFolder/ChangeRegistrationByParentFolderTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\FavoriteVideo\ParentFolder;
 
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Testing\TestResponse;
 use Tests\TestCase;
-use App\Models\User;
 
 class ChangeRegistrationByParentFolderTest extends TestCase
 {
