+++ backend/tests/Feature/FavoriteVideo/ParentFolder/FetchParentFoldersTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\FavoriteVideo\ParentFolder;
 
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
-use App\Models\User;
 
 class FetchParentFoldersTest extends TestCase
 {
