+++ backend/tests/Feature/FavoriteVideo/ParentFolder/StoreParentFolderTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\FavoriteVideo\ParentFolder;
 
 use App\Models\ParentFolder;
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
-use App\Models\User;
 
 class StoreParentFolderTest extends TestCase
 {
