+++ backend/tests/Feature/FavoriteVideo/ParentFolder/UpdateParentFolderTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\FavoriteVideo\ParentFolder;
 
 use App\Models\ParentFolder;
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Testing\TestResponse;
 use Tests\TestCase;
-use App\Models\User;
 
 class UpdateParentFolderTest extends TestCase
 {
